+++ api/models/User.js
@@ -58,17 +58,27 @@ var User = {
         next();
     },
     afterDestroy: function (users, next) {
+        if(users.length == 0){
+            return next();
+        }
         users.forEach(function (user, i) {
             //TODO clean all datas
+            sails.log.info('destroy user ' + user.id);
             sails.models.profile.destroy({
                 id: user.profile
             }).exec(function (err) {
-
+                if(err){
+                    return next(err);
+                }
+                sails.log.info('destroy profile');
                 //Remove user fav list
-                sails.models.userFavList.destroy({
+                UserFavList.destroy({
                     owner: user.id
                 }).exec(function (err) {
-
+                    if(err){
+                        return next(err);
+                    }
+                    sails.log.info('destroy fav list');
                     //Remove user conversations
                     sails.models.conversation.destroy({
                         $or: [
@@ -76,11 +86,18 @@ var User = {
                             {interlocutor: user.id}
                         ]
                     }).exec(function (err) {
-
+                        if(err){
+                            sails.log.info('error destroying convs');
+                            return next(err);
+                        }
+                        sails.log.info('destroy convs');
                         //Remove user diplomas
                         sails.models.diploma.destroy({
                             owner: user.id
                         }).exec(function (err) {
+                            if(err){
+                                return next(err);
+                            }
                             next();
                         });
                     });
